feat: renamed devConsole.js to smsClient.js, also updated on all relevant places
