+++ src/commands/api-mesh/__tests__/get.test.js
@@ -23,7 +23,7 @@ jest.mock('../../../helpers', () => ({
 	initSdk: jest.fn().mockResolvedValue({}),
 	initRequestId: jest.fn().mockResolvedValue({}),
 }));
-jest.mock('../../../lib/devConsole');
+jest.mock('../../../lib/smsClient');
 
 const mockConsoleCLIInstance = {};
 const selectedOrg = { id: '1234', code: 'CODE1234@AdobeOrg', name: 'ORG01', type: 'entp' };
@@ -34,7 +34,7 @@ const { writeFile } = require('fs/promises');
 const { initSdk, initRequestId } = require('../../../helpers');
 const GetCommand = require('../get');
 const mockGetMeshConfig = require('../../__fixtures__/sample_mesh.json');
-const { getMeshId, getMesh } = require('../../../lib/devConsole');
+const { getMeshId, getMesh } = require('../../../lib/smsClient');
 
 let logSpy = null;
 let errorLogSpy = null;
