+++ src/commands/api-mesh/__tests__/log-get.test.js
@@ -24,7 +24,7 @@ jest.mock('../../../helpers', () => ({
 	initRequestId: jest.fn().mockResolvedValue({}),
 }));
 
-jest.mock('../../../lib/devConsole');
+jest.mock('../../../lib/smsClient');
 
 const mockConsoleCLIInstance = {};
 const selectedOrg = { id: '1234', code: 'CODE1234@AdobeOrg', name: 'ORG01', type: 'entp' };
@@ -34,7 +34,7 @@ const selectedWorkspace = { id: '123456789', title: 'Workspace01' };
 const { writeFile } = require('fs/promises');
 const { initSdk } = require('../../../helpers');
 const FetchLogsCommand = require('../log-get');
-const { getLogsByRayId, getMeshId } = require('../../../lib/devConsole');
+const { getLogsByRayId, getMeshId } = require('../../../lib/smsClient');
 const os = require('os');
 let logSpy = null;
 let errorLogSpy = null;
