+++ src/commands/api-mesh/create.js
@@ -26,7 +26,7 @@ const {
 	validateSecretsFile,
 	encryptSecrets,
 } = require('../../utils');
-const { createMesh, getPublicEncryptionKey } = require('../../lib/devConsole');
+const { createMesh, getPublicEncryptionKey } = require('../../lib/smsClient');
 const { buildMeshUrl } = require('../../urlBuilder');
 
 class CreateCommand extends Command {
