+++ src/commands/api-mesh/delete.js
@@ -14,7 +14,7 @@ const { Command } = require('@oclif/command');
 const logger = require('../../classes/logger');
 const { initSdk, initRequestId, promptConfirm } = require('../../helpers');
 const { ignoreCacheFlag, autoConfirmActionFlag } = require('../../utils');
-const { getMeshId, deleteMesh } = require('../../lib/devConsole');
+const { getMeshId, deleteMesh } = require('../../lib/smsClient');
 
 require('dotenv').config();
