+++ src/commands/api-mesh/log-get-bulk.js
@@ -2,7 +2,7 @@ const { Command } = require('@oclif/core');
 const path = require('path');
 const fs = require('fs');
 const { initRequestId, initSdk, promptConfirm } = require('../../helpers');
-const { getMeshId, getPresignedUrls } = require('../../lib/devConsole');
+const { getMeshId, getPresignedUrls } = require('../../lib/smsClient');
 const logger = require('../../classes/logger');
 const axios = require('axios');
 const {
