+++ src/commands/api-mesh/log-list.js
@@ -3,7 +3,7 @@ const { Command } = require('@oclif/core');
 const logger = require('../../classes/logger');
 const { initSdk, initRequestId } = require('../../helpers');
 const { ignoreCacheFlag, fileNameFlag } = require('../../utils');
-const { getMeshId, listLogs } = require('../../lib/devConsole');
+const { getMeshId, listLogs } = require('../../lib/smsClient');
 const { appendFileSync, existsSync } = require('fs');
 const { ux } = require('@oclif/core/lib/cli-ux');
 const path = require('path');
