+++ src/commands/api-mesh/run.js
@@ -35,7 +35,7 @@ const {
 	writeSecretsFile,
 } = require('../../helpers');
 const logger = require('../../classes/logger');
-const { getMeshId, getMeshArtifact } = require('../../lib/devConsole');
+const { getMeshId, getMeshArtifact } = require('../../lib/smsClient');
 require('dotenv').config();
 const { runServer } = require('../../server');
 const { fixPlugins } = require('../../fixPlugins');
